refactor: refatora o UserController e o TaskController para usar o ResponseEntity e enviar mensagens de erros mais agradáveis ao usuário

+++ src/main/java/dev/viniciusmacedo/task_manager/tasks/TaskController.java
@@ -1,5 +1,7 @@
 package dev.viniciusmacedo.task_manager.tasks;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -15,28 +17,46 @@ public class TaskController {
     }
 
     @GetMapping("/{id}")
-    public TaskDTO getTaskById(@PathVariable Long id) {
-        return taskService.findById(id);
+    public ResponseEntity<?> getTaskById(@PathVariable Long id) {
+        if (taskService.findById(id) != null) {
+            return ResponseEntity.ok(taskService.findById(id));
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Tarefa não encontrada");
+        }
     }
 
     @GetMapping()
-    public List<TaskDTO> findAll() {
-        return taskService.findAll();
+    public ResponseEntity<?> findAll() {
+        if (taskService.findAll().isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Tarefa não encontrada");
+        } else {
+            return ResponseEntity.ok(taskService.findAll());
+
+        }
     }
 
     @PutMapping("/{id}")
-    public TaskDTO editTask(@PathVariable Long id, @RequestBody TaskDTO taskDTO) {
-        return taskService.editById(id, taskDTO);
+    public ResponseEntity<?> editTask(@PathVariable Long id, @RequestBody TaskDTO taskDTO) {
+        if (taskService.findById(id) != null) {
+            return ResponseEntity.ok(taskService.editById(id, taskDTO));
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Tarefa não encontrada");
+        }
     }
 
     @PostMapping
-    public TaskDTO createTask(@RequestBody TaskDTO task) {
-        return taskService.createTask(task);
+    public ResponseEntity<TaskDTO> createTask(@RequestBody TaskDTO task) {
+        return ResponseEntity.status(HttpStatus.CREATED).body(taskService.createTask(task));
     }
 
     @DeleteMapping("/{id}")
-    public void deleteTask(@PathVariable Long id) {
-        taskService.deleteById(id);
+    public ResponseEntity<?> deleteTask(@PathVariable Long id) {
+        if (taskService.findById(id) != null) {
+            taskService.deleteById(id);
+            return ResponseEntity.noContent().build();
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Tarefa não encontrada");
+        }
     }
