+++ src/main/java/dev/viniciusmacedo/task_manager/users/UserController.java
@@ -1,5 +1,7 @@
 package dev.viniciusmacedo.task_manager.users;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -15,27 +17,45 @@ public class UserController {
     }
 
     @GetMapping("/{id}")
-    public UserDTO findUserById(@PathVariable Long id) {
-        return userService.findById(id);
+    public ResponseEntity<?> findUserById(@PathVariable Long id) {
+        if (userService.findById(id) != null) {
+            return ResponseEntity.ok(userService.findById(id));
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Usuário de id: " + id + " não encontrado");
+        }
+
     }
 
     @GetMapping
-    public List<UserDTO> findAll() {
-        return userService.findAll();
+    public ResponseEntity<?> findAll() {
+        if (userService.findAll() != null) {
+            return ResponseEntity.ok(userService.findAll());
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Nenhum usuário cadastrado");
+        }
     }
 
     @PutMapping("{id}")
-    public UserDTO editById(@PathVariable Long id, @RequestBody UserDTO userDTO) {
-        return userService.editById(id, userDTO);
+    public ResponseEntity<?> editById(@PathVariable Long id, @RequestBody UserDTO userDTO) {
+        if (userService.findById(id) != null) {
+            return ResponseEntity.ok(userService.editById(id, userDTO));
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Usuário de id: " + id + " não encontrado");
+        }
     }
 
     @PostMapping()
-    public UserDTO createUser(@RequestBody UserDTO user) {
-        return userService.createUser(user);
+    public ResponseEntity<UserDTO> createUser(@RequestBody UserDTO user) {
+        return ResponseEntity.status(HttpStatus.CREATED).body(userService.createUser(user));
     }
 
     @DeleteMapping("/{id}")
-    public void deleteUser(@PathVariable Long id) {
-        userService.deleteById(id);
+    public ResponseEntity<?> deleteUser(@PathVariable Long id) {
+        if (userService.findById(id) != null) {
+            userService.deleteById(id);
+            return ResponseEntity.noContent().build();
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Usuário de id: " + id + " não encontrado");
+        }
     }
 }
